refactor: move config to mixin
