+++ lib/base_loader.js
@@ -7,6 +7,7 @@ const isFunction = require('is-type-of').function;
 const debug = require('debug')('egg-loader');
 const Loader = require('./loader');
 const loadFile = require('./utils').loadFile;
+const getHomedir = require('./utils').getHomedir;
 const Emitter = require('events').EventEmitter;
 
 class EggLoader {
@@ -46,6 +47,14 @@ class EggLoader {
      */
     this.serverEnv = this.getServerEnv();
     debug('Loaded serverEnv %j', this.serverEnv);
+
+    this.appInfo = {
+      name: this.getAppname(),
+      baseDir: this.options.baseDir,
+      env: this.serverEnv,
+      HOME: getHomedir(),
+      pkg: this.pkg,
+    };
   }
 
   /**
@@ -237,7 +246,7 @@ class EggLoader {
  */
 const loaders = [
   require('./plugin_loader'),
-  require('./config_loader'),
+  require('./mixin/config'),
   require('./mixin/extend'),
   require('./mixin/custom'),
   require('./proxy_loader'),
