+++ lib/mixin/middleware.js
@@ -2,7 +2,7 @@
 
 const join = require('path').join;
 const is = require('is-type-of');
-const debug = require('debug')('egg:loader:middleware');
+const debug = require('debug')('egg-loader:middleware');
 const inspect = require('util').inspect;
 
 module.exports = {
